Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/strict-origin-when-cross-origin/attr-referrer/same-origin/http-https/img-tag/keep-origin-redirect/upgrade-protocol.http-expected.txt



FAIL The referrer URL is origin when a
                                 document served over http requires an https
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/strict-origin-when-cross-origin/attr-referrer/same-origin/http-https/img-tag/keep-origin-redirect/upgrade-protocol.http.html"
FAIL The referrer URL is origin when a
                                 document served over http requires an https
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is same-origin. (srcdoc iframe inherits parent) assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/strict-origin-when-cross-origin/attr-referrer/same-origin/http-https/img-tag/keep-origin-redirect/upgrade-protocol.http.html"
PASS The referrer URL is origin when a
                                 document served over http requires an https
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is same-origin. (overridden by srcdoc iframe)
